Moved access token endpoints functions in properly routed controllers
