+++ backend/exceptions/auth/InvalidTokenTypeError.js
@@ -1,0 +1,3 @@
+class InvalidTokenTypeError extends Error {}
+
+module.exports = InvalidTokenTypeError;
