+++ backend/server.js
@@ -212,185 +212,6 @@ app.get('/', async (req, res) => {
     res.send('lequiz.io-backend container');
 });
 
-/**
- * AUTH JWT - TODO ROUTER & FILES
- */
-
-class InvalidTokenTypeError extends Error {}
-
-const jwtSecret = 'superSecret'; // TODO config
-
-const refreshTokens = []; // TODO store in database
-
-/**
- * Verifies a JWT token, and optionnaly its type
- * @param token string
- * @param type string|string[] (optional) if set, the token must be of that type (accessToken, refreshToken)
- * @return {{
- *     verified: boolean
- *     payload: object
- *     error: string
- * }}
- */
-const verifyToken = (token, type = null) => {
-    const result = {};
-
-    try {
-        const payload = jwt.verify(token, jwtSecret);
-        if(type !== null) { // Verify token type (accessToken/refreshToken)
-            verifyTokenType(payload, type);
-        }
-
-        result.verified = true;
-        result.payload = payload;
-        return result;
-    } catch(e) {
-        result.verified = false;
-        switch(e.constructor.name) {
-            case 'JsonWebTokenError':
-                result.error = 'Malformed token';
-                break;
-            case 'TokenExpiredError':
-                result.error = 'Token has expired';
-                break;
-            case 'NotBeforeError':
-                result.error = 'Token is not yet valid';
-                break;
-            case 'InvalidTokenTypeError':
-                result.error = e.message;
-                break;
-            default:
-                throw e;
-        }
-
-        return result;
-    }
-}
-
-/**
- * Verifies the type contained in a JWT token payload
- * @param tokenPayload object
- * @param expectedType string|string[]
- * @throw InvalidTokenTypeError if token type does not match
- */
-const verifyTokenType = (tokenPayload, expectedType) => {
-    if(!tokenPayload.hasOwnProperty('type')) {
-        throw new InvalidTokenTypeError('Token type not found in payload');
-    }
-
-    if(typeof expectedType === 'string') {
-        expectedType = [expectedType];
-    }
-
-    if(!expectedType.includes(tokenPayload.type)) {
-        throw new InvalidTokenTypeError(`Type ${tokenPayload.type} does not match any expected type (${expectedType.join(', ')})`);
-    }
-}
-
-const generateToken = (type, initialPayload = {}) => {
-    let expiresIn;
-    switch(type) {
-        case 'accessToken':
-            expiresIn = 3;
-            break;
-        case 'refreshToken':
-            expiresIn = 6;
-            break;
-        default:
-            throw new InvalidTokenTypeError();
-    }
-
-    const payload = {...initialPayload};
-    delete payload.iat;
-    delete payload.exp;
-    delete payload.type;
-
-    payload.type = type;
-
-    return jwt.sign(payload, jwtSecret, {
-        expiresIn,
-    });
-}
-
-const invalidateRefreshToken = (refreshToken) => {
-    const index = refreshTokens.indexOf(refreshToken);
-    if(index > -1) refreshTokens.splice(index, 1);
-}
-
-const saveRefreshToken = (refreshToken) => {
-    refreshTokens.push(refreshToken);
-}
-
-app.get('/auth/verify-token', (req, res) => {
-    const token = req.headers.authorization;
-    res.status(200);
-    const response = {
-        valid: true,
-    };
-    const verification = verifyToken(token, ['accessToken', 'refreshToken']);
-    if(verification.verified) {
-        response.valid = true;
-        response.type = verification.payload.type;
-    } else {
-        response.valid = false;
-        response.error = verification.error;
-    }
-
-    res.send(response);
-});
-
-app.get('/auth/access-token', (req, res) => {
-    let fromRefreshToken = false;
-    if(req.query.refreshToken !== undefined) {
-        fromRefreshToken = true;
-    }
-
-    if(fromRefreshToken) {
-        const inputRefreshToken = req.query.refreshToken;
-        const response = {};
-
-        if(!refreshTokens.includes(inputRefreshToken)) {
-            res.status(400);
-            response.error = 'Unknown refresh token';
-            res.send(response);
-            return;
-        }
-
-        const verification = verifyToken(inputRefreshToken, 'refreshToken');
-        if(!verification.verified) {
-            response.error = verification.error
-            res.status(400);
-            res.send(response);
-            return;
-        }
-
-        const refreshTokenPayload = verification.payload;
-
-        newAccessToken = generateToken('accessToken', refreshTokenPayload);
-        newRefreshToken = generateToken('refreshToken', refreshTokenPayload);
-
-        invalidateRefreshToken(inputRefreshToken);
-        saveRefreshToken(newRefreshToken);
-
-        response.accessToken = newAccessToken;
-        response.refreshToken = newRefreshToken;
-        res.send(response);
-        return;
-    }
-
-    const fakeUserId = Math.ceil(Math.random() * 10000);
-
-    const accessToken = generateToken('accessToken', {userId: fakeUserId});
-    const refreshToken = generateToken('refreshToken', {userId: fakeUserId});
-
-    refreshTokens.push(refreshToken);
-
-    res.send({
-        'accessToken': accessToken,
-        'refreshToken': refreshToken,
-    })
-})
-
 app.listen(port,() => {
     console.log(`Server running at http://localhost:${port}`);
 });
